requirements.txt
Pyqt5==5.15.5
mysql-connector-python==9.0.0